fairchain-ai/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.2
imbalanced-learn==0.12.3
shap==0.46.0
google-genai==1.7.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
pydantic==2.9.2
reportlab==4.2.2
google-generativeai==0.8.3
httpx==0.27.2